fix(env): revert lock file
